requirements.txt
requests
sys
pillow
json
tkinter
io